Move measure creation from measure data  to distinct file
